Add a general event_display_data function
This collapses Imprint-specific functions into Event itself,
since there was actually no need to restrict any of them to
a separate model.

The general event_display_data is intended for related resources,
which don't care about splitting out event notes and just want
a concise description of events that respects displayLabel.

Fixes an issue with records like cj555pv1585 that were not
displaying all of the event information.
