+++ lib/cocina_display/related_resource.rb
@@ -59,8 +59,7 @@ module CocinaDisplay
     def display_data
       title_display_data +
         contributor_display_data +
-        event_date_display_data +
-        event_note_display_data +
+        event_display_data +
         general_note_display_data +
         preferred_citation_display_data +
         access_display_data +
